Computes the CMPS score of a comparison between two bullet profiles/signatures

Compute the Congruent Matching Profile Segments (CMPS) score based on two bullet profiles/signatures. The reference profile will be divided into consecutive, non-overlapping, basis segments of the same length. Then the number of segments that are congruent matching will be found as the CMPS score. By default, extract_feature_cmps implements the algorithm with multi-peak insepction at three different segment scales. By setting npeaks.set as a single-length vector, users can switch to the algorithm with multi-peak inspection at the basis scale only.

extract_feature_cmps(
  x,
  y,
  seg_length = 50,
  Tx = 25,
  npeaks.set = c(5, 3, 1),
  include = NULL
)

Arguments

x
a numeric vector, vector of the reference bullet signature/profile that will be divided into basis segments

y
a numeric vector, vector of the comparison bullet signature/profile

seg_length
integer, the length of a basis segment

Tx
integer, the tolerance zone is +/- Tx

npeaks.set
a numeric vector, specify the number of peaks to be found for each different scale of segment. The number of peaks for the highest segment scale must be 1, i.e., the very last element of npeaks.set must be 1.

If length(npeaks.set) == 1, the algorithm uses multi-peak inspection only at the basis scale;

If length(npeaks.set) > 1, the algorithm uses multi-peak inspection at different segment scales.

By default, npeaks.set = c(5,3,1). Including more segment scales will reduce the number of false positive results

include
NULL or a vector of character strings indicating what additional information should be included in the output of extract_feature_cmps. All possible options are: "nseg", "congruent.pos", "congruent.seg", "congruent.seg.idx", "pos.df", "ccp.list","segments", and "parameters". If one wants to include them all, one can use include = "full_result

Value

a numeric value or a list

if include = NULL, returns the CMPS score (a numeric value) only

if include = one or a vector of strings listed above:

nseg: number of basis segments

congruent.seg: a vector of boolean values. TRUE means this basis segment is a congruent matching profile segment (CMPS)

congruent.seg.idx: the indices of all CMPS

pos.df: a dataframe that includes positions of correlation peaks and the CMPS score of these positions

ccp.list: a list of consistent correlation peaks of each basis segment.

segments: a list of all basis segments

parameters: a list that stores all parameters used in the function call

References

Chen, Zhe, Wei Chu, Johannes A Soons, Robert M Thompson, John Song, and Xuezeng Zhao. 2019. “Fired Bullet Signature Correlation Using the Congruent Matching Profile Segments (CMPS) Method.” Forensic Science International, December, #109964. https://doi.org/10.1016/j.forsciint.2019.109964.

Examples

library(bulletxtrctr)
library(tidyverse)
library(x3ptools)
library(CMPS)

data("bullets")
land2_3 <- bullets$sigs[bullets$bulletland == "2-3"][[1]]
land1_2 <- bullets$sigs[bullets$bulletland == "1-2"][[1]]

# compute cmps

# algorithm with multi-peak insepction at three different segment scales
cmps_with_multi_scale <- extract_feature_cmps(land2_3$sig, land1_2$sig, include = "full_result" )

# algorithm with multi-peak inspection at the basis scale only
cmps_without_multi_scale <- extract_feature_cmps(land2_3$sig, land1_2$sig, 
                                                 npeaks.set = 5, include = "full_result" )
if (FALSE) {
library(tidyverse)
library(bulletxtrctr)

lands <- unique(bullets$bulletland)

comparisons <- data.frame(expand.grid(land1 = lands[1:6], land2 = lands[7:12]), 
                          stringsAsFactors = FALSE)

comparisons <- comparisons %>% mutate(
  aligned = purrr::map2(.x = land1, .y = land2, 
                        .f = function(xx, yy) {
                          land1 <- bullets$sigs[bullets$bulletland == xx][[1]]
                          land2 <- bullets$sigs[bullets$bulletland == yy][[1]]
                          land1$bullet <- "first-land"
                          land2$bullet <- "second-land"
                          
                          sig_align(land1$sig, land2$sig)
                        }))

comparisons <- comparisons %>% 
  mutate(cmps = aligned %>% purrr::map(.f = function(a) {
    extract_feature_cmps(a$lands$sig1, a$lands$sig2, include = "nseg")
  }))

# comparisons.cmps <- comparisons %>% 
#   mutate(cmps = aligned %>% purrr::map_dbl(.f = function(a) {
#     extract_feature_cmps(a$lands$sig1, a$lands$sig2, include = NULL)
#   }))
# comparisons.cmps %>% select(land1, land2, cmps) 

comparisons <- comparisons %>% 
  mutate(
    cmps_score = sapply(comparisons$cmps, function(x) x$CMPS.score),
    cmps_nseg = sapply(comparisons$cmps, function(x) x$nseg)
  )

cp1 <- comparisons %>% select(land1, land2, cmps_score, cmps_nseg)
cp1
}