Computes the CMPS score of a comparison between two bullet profiles/signatures

Compute the Congruent Matching Profile Segments (CMPS) score based on two bullet profiles/signatures. The reference profile will be divided into consecutive, non-overlapping, basis segments of the same length. Then the number of segments that are congruent matching will be found as the CMPS score. By default, extract_feature_cmps implements the algorithm with multi-peak insepction at three different segment scales. By setting seg_scale_max = 1, users can switch to the algorithm with multi-peak inspection at the basis scale only.

extract_feature_cmps(
  x,
  y,
  seg_length = 50,
  seg_scale_max = 3,
  Tx = 25,
  npeaks.set = c(5, 3, 1),
  full_result = FALSE
)

Arguments

x
a numeric vector, vector of the reference bullet signature/profile that will be divided into basis segments

y
a numeric vector, vector of the comparison bullet signature/profile

seg_length
integer, integer, the length of a basis segment

seg_scale_max
integer, the maximum scale of the segment.

If setting seg_scale_max = 1, the algorithm uses multi-peak inspection only at the basis scale;

If setting seg_scale_max to be an integer greater than 1, the algorithm uses multi-peak inspection at different segment scales.

By default, seg_scale_max = 3. Increasing seg_scale_max will reduce the number of false positive results

Tx
integer, the tolerance zone is +/- Tx

npeaks.set
a numeric vector, specify the number of peaks to be found for each different scale of segment. The number of peaks for the highest segment scale must be 1, i.e., the very last element of npeaks.set must be 1.

full_result
boolean, whether or not to return the full CMPS result

Value

integer or a list

if full_result = FALSE, return the CMPS number only

if full_result = TRUE, return a list of four elements:

CMPS.score: computed CMPS score

rec.position: the recommended position that results in the CMPS score

pos.df: a dataframe that includes all positions and their corresponding CMPS score

nseg: the number of basis segments obtained from the reference profile

Examples

library(bulletxtrctr)
library(tidyverse)

#> -- Attaching packages --------------------------------------------------------------------- tidyverse 1.3.0 --

#> v ggplot2 3.3.0     v purrr   0.3.3
#> v tibble  2.1.3     v dplyr   0.8.5
#> v tidyr   1.0.2     v stringr 1.4.0
#> v readr   1.3.1     v forcats 0.5.0

#> -- Conflicts ------------------------------------------------------------------------ tidyverse_conflicts() --
#> x dplyr::filter() masks stats::filter()
#> x dplyr::lag()    masks stats::lag()

library(x3ptools)
library(CMPS)

data("bullets")
land2_3 <- bullets$sigs[bullets$bulletland == "2-3"][[1]]
land1_2 <- bullets$sigs[bullets$bulletland == "1-2"][[1]]

# compute cmps

# algorithm with multi-peak insepction at three different segment scales
cmps_with_multi_scale <- extract_feature_cmps(land2_3$sig, land1_2$sig, full_result = TRUE)

# algorithm with multi-peak inspection at the basis scale only
cmps_without_multi_scale <- extract_feature_cmps(land2_3$sig, land1_2$sig, seg_scale_max = 1,
                                                 npeaks.set = 5, full_result = TRUE)